An R6 Class object, a collection of model for population of sbm netowrks

An R6 Class object, a collection of model for population of sbm netowrks

An R6 Class object, a collection of model for population of sbm netowrks

Methods

Public methods

bmpop$plot()

bmpop$new()

bmpop$optimize_sbm()

bmpop$optimize_from_sbm()

bmpop$optimize_spectral()

bmpop$optimize_init()

bmpop$optimize_from_zinit()

bmpop$burn_in()

bmpop$forward_pass()

bmpop$backward_pass()

bmpop$optimize()

bmpop$forward_backward()

bmpop$choose_models()

bmpop$show()

bmpop$print()

bmpop$clone()

Method plot()

Usage

bmpop$plot(type = "trace", ...)

Method new()

Usage

bmpop$new(
  netlist = NULL,
  net_id = NULL,
  directed = NULL,
  model = "bernoulli",
  free_density = FALSE,
  free_mixture = FALSE,
  fit_sbm = NULL,
  Z_init = NULL,
  global_opts = list(),
  fit_opts = list()
)

Method optimize_sbm()

Usage

bmpop$optimize_sbm()

Method optimize_from_sbm()

Usage

bmpop$optimize_from_sbm(index, Q, nb_clusters)

Method optimize_spectral()

Usage

bmpop$optimize_spectral(index, Q, nb_clusters)

Method optimize_init()

Usage

bmpop$optimize_init(index, Z, Q, nb_clusters, Cpi = NULL, Calpha = NULL)

Method optimize_from_zinit()

Usage

bmpop$optimize_from_zinit(index, Q, nb_clusters)

Method burn_in()

Usage

bmpop$burn_in()

Method forward_pass()

Usage

bmpop$forward_pass(
  Q_min = self$global_opts$Q_min,
  Q_max = self$global_opts$Q_max,
  index = seq(self$M),
  nb_clusters = 1L
)

Method backward_pass()

Usage

bmpop$backward_pass(
  Q_min = self$global_opts$Q_min,
  Q_max = self$global_opts$Q_max,
  index = seq(self$M),
  nb_clusters = 1L
)

Method optimize()

Usage

bmpop$optimize()

Method forward_backward()

Usage

bmpop$forward_backward()

Method choose_models()

Usage

bmpop$choose_models(models, Q, index = seq(self$M), nb_clusters = 1L)

Method show()

Usage

bmpop$show(type = "Fitted Collection of Simple SBM")

Method print()

Usage

bmpop$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

bmpop$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.