An R6 Class object, a fitted population of netowrks sbm once $optimize() is done

An R6 Class object, a fitted population of netowrks sbm once $optimize() is done

An R6 Class object, a fitted population of netowrks sbm once $optimize() is done

Methods

Public methods

fitBipartiteSBMPop$new()

fitBipartiteSBMPop$compute_map()

fitBipartiteSBMPop$objective()

fitBipartiteSBMPop$vb_tau_alpha()

fitBipartiteSBMPop$vb_tau_pi()

fitBipartiteSBMPop$entropy_tau()

fitBipartiteSBMPop$fn_vb_alpha_delta()

fitBipartiteSBMPop$gr_vb_alpha_delta()

fitBipartiteSBMPop$eval_g0_vb_alpha_delta()

fitBipartiteSBMPop$eval_jac_g0_vb_alpha_delta()

fitBipartiteSBMPop$update_alpha_delta()

fitBipartiteSBMPop$compute_vbound()

fitBipartiteSBMPop$compute_penalty()

fitBipartiteSBMPop$compute_icl()

fitBipartiteSBMPop$compute_icl_clustering()

fitBipartiteSBMPop$compute_exact_icl()

fitBipartiteSBMPop$compute_exact_icl_iid()

fitBipartiteSBMPop$update_map_parameters()

fitBipartiteSBMPop$fixed_point_tau()

fitBipartiteSBMPop$fixed_point_alpha_delta()

fitBipartiteSBMPop$update_pi()

fitBipartiteSBMPop$update_alpham()

fitBipartiteSBMPop$update_alpha()

fitBipartiteSBMPop$init_clust()

fitBipartiteSBMPop$make_permutation()

fitBipartiteSBMPop$m_step()

fitBipartiteSBMPop$ve_step()

fitBipartiteSBMPop$update_mqr()

fitBipartiteSBMPop$optimize()

fitBipartiteSBMPop$clone()

Method new()

Usage

fitBipartiteSBMPop$new(
  A = NULL,
  Q = NULL,
  Z = NULL,
  mask = NULL,
  net_id = NULL,
  model = "bernoulli",
  free_mixture = TRUE,
  free_density = TRUE,
  directed = NULL,
  init_method = "spectral",
  weight = NULL,
  fit_opts = list(algo_ve = "fp", approx_pois = TRUE, minibatch = TRUE, verbosity = 1)
)

Method compute_map()

Usage

fitBipartiteSBMPop$compute_map()

Method objective()

Usage

fitBipartiteSBMPop$objective()

Method vb_tau_alpha()

Usage

fitBipartiteSBMPop$vb_tau_alpha(m, map = FALSE)

Method vb_tau_pi()

Usage

fitBipartiteSBMPop$vb_tau_pi(m, map = FALSE)

Method entropy_tau()

Usage

fitBipartiteSBMPop$entropy_tau(m)

Method fn_vb_alpha_delta()

Usage

fitBipartiteSBMPop$fn_vb_alpha_delta(par, emqr, nmqr)

Method gr_vb_alpha_delta()

Usage

fitBipartiteSBMPop$gr_vb_alpha_delta(par, emqr, nmqr)

Method eval_g0_vb_alpha_delta()

Usage

fitBipartiteSBMPop$eval_g0_vb_alpha_delta(par, emqr, nmqr)

Method eval_jac_g0_vb_alpha_delta()

Usage

fitBipartiteSBMPop$eval_jac_g0_vb_alpha_delta(par, emqr, nmqr)

Method update_alpha_delta()

Usage

fitBipartiteSBMPop$update_alpha_delta(map = FALSE)

Method compute_vbound()

Usage

fitBipartiteSBMPop$compute_vbound()

Method compute_penalty()

Usage

fitBipartiteSBMPop$compute_penalty()

Method compute_icl()

Usage

fitBipartiteSBMPop$compute_icl(map = FALSE)

Method compute_icl_clustering()

Usage

fitBipartiteSBMPop$compute_icl_clustering(map = TRUE)

Method compute_exact_icl()

Usage

fitBipartiteSBMPop$compute_exact_icl()

Method compute_exact_icl_iid()

Usage

fitBipartiteSBMPop$compute_exact_icl_iid()

Method update_map_parameters()

Usage

fitBipartiteSBMPop$update_map_parameters()

Method fixed_point_tau()

Usage

fitBipartiteSBMPop$fixed_point_tau(m, d, max_iter = 10, tol = 0.001)

Method fixed_point_alpha_delta()

Usage

fitBipartiteSBMPop$fixed_point_alpha_delta(
  map = FALSE,
  max_iter = 50,
  tol = 1e-06
)

Method update_pi()

Usage

fitBipartiteSBMPop$update_pi(m, map = FALSE)

Method update_alpham()

Usage

fitBipartiteSBMPop$update_alpham(m, map = FALSE)

Method update_alpha()

Usage

fitBipartiteSBMPop$update_alpha(map = FALSE)

Method init_clust()

Usage

fitBipartiteSBMPop$init_clust()

Method make_permutation()

Usage

fitBipartiteSBMPop$make_permutation()

Method m_step()

Usage

fitBipartiteSBMPop$m_step(map = FALSE, max_iter = 100, tol = 0.001, ...)

Method ve_step()

Usage

fitBipartiteSBMPop$ve_step(m, max_iter = 20, tol = 0.001, ...)

Method update_mqr()

Usage

fitBipartiteSBMPop$update_mqr(m)

Method optimize()

Usage

fitBipartiteSBMPop$optimize(max_step = 100, tol = 0.001, ...)

Method clone()

The objects of this class are cloneable with this method.

Usage

fitBipartiteSBMPop$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.